Added movement and other stuff
Reworked advantage system, added camera movement/looking (may need even more adjustment)

Player can now attack the enemy, including running up on them. Will still need to add the attacking animation, though.

Player now will run up to enemy, and camera will follow.

Changed attack types to shorts.

Updated Unit reads on their weaknesses through using the ScriptableObjects attached.

Added party identification to Awake() in Battle Systems for testing purposes with advantage reading, will need to be removed later.

Updated

Note:
Will need to add an "in-battle" variable in the future so more enemies can join the battle.
PCC playerSpeed is now public, and should be set to around 7 for speedy movement.

Known bugs:
Target perspective bug is still in effect.
Damage is still calculated, applied, and shown before the player would even begin their attack.
The Select.Play() command doesn't seem to work in the Update function for some reason.... will need to try to move it to playerAttack() coroutine maybe?
Killing 1 enemy for some reason will immediately make the players win.
WILL NEED TO REMOVE THE PARTY CREATION COMMAND FROM THE AWAKE FUNCTION TO PREVENT MULTIPLE PARTIES BEING CREATED

+++ Assets/Scripts/Digital World/ShadowBase.cs
@@ -13,7 +13,7 @@ public class ShadowBase : ScriptableObject
     public int maxHP, maxSP;
     public int str, mag, en, ag, lu;
     public int weapon, armor;
-    public int[] weak, resist, absorb, reflect;    //0=weak, 1=normal, 2=strong, 3=reflect, 4=absorb
+    public short[] weak, resist, absorb, reflect;    //0=weak, 1=normal, 2=strong, 3=reflect, 4=absorb
     //P=0, G=1, F=2, I=3, L=4, W=5, Ps=6, N=7, B=8, C=9, A=10
